# ISPN-5451 data container striping