--- a/client/hotrod-client/src/main/java/org/infinispan/client/hotrod/impl/protocol/Codec27.java
+++ b/client/hotrod-client/src/main/java/org/infinispan/client/hotrod/impl/protocol/Codec27.java
@@ -3,7 +3,7 @@ package org.infinispan.client.hotrod.impl.protocol;
 import org.infinispan.client.hotrod.RemoteCache;
 import org.infinispan.client.hotrod.impl.operations.OperationsFactory;
 import org.infinispan.commons.util.CloseableIterator;
-import org.infinispan.commons.util.CloseableIteratorMapper;
+import org.infinispan.commons.util.IteratorMapper;
 
 import io.netty.buffer.ByteBuf;
 
@@ -20,7 +20,7 @@ public class Codec27 extends Codec26 {
    @Override
    public <K> CloseableIterator<K> keyIterator(RemoteCache<K, ?> remoteCache, OperationsFactory operationsFactory,
          int batchSize) {
-      return new CloseableIteratorMapper<>(remoteCache.retrieveEntries(
+      return new IteratorMapper<>(remoteCache.retrieveEntries(
             // Use the ToEmptyBytesKeyValueFilterConverter to remove value payload
             "org.infinispan.server.hotrod.HotRodServer$ToEmptyBytesKeyValueFilterConverter", batchSize),
             e -> (K) e.getKey());