--- a/commons/src/main/java/org/infinispan/commons/util/IteratorMapper.java
+++ b/commons/src/main/java/org/infinispan/commons/util/IteratorMapper.java
@@ -5,11 +5,12 @@ import java.util.function.Function;
 
 /**
  * A iterator that maps each value to the output of the Function.  Note that the remove is supported if the iterator
- * originally supported remove.
+ * originally supported remove. This iterator implements {@link CloseableIterator} and will close the provided iterator
+ * if it also implemented CloseableIterator.
  * @author William Burns
  * @since 8.0
  */
-public class IteratorMapper<E, S> implements Iterator<S> {
+public class IteratorMapper<E, S> implements CloseableIterator<S> {
    private final Iterator<? extends E> iterator;
    private final Function<? super E, ? extends S> function;
 
@@ -35,5 +36,12 @@ public class IteratorMapper<E, S> implements Iterator<S> {
    @Override
    public void remove() {
       iterator.remove();
+   }
+
+   @Override
+   public void close() {
+      if (iterator instanceof CloseableIterator) {
+         ((CloseableIterator) iterator).close();
+      }
    }
 }