--- a/core/src/main/java/org/infinispan/atomic/impl/AtomicKeySetImpl.java
+++ b/core/src/main/java/org/infinispan/atomic/impl/AtomicKeySetImpl.java
@@ -13,7 +13,7 @@ import java.util.function.Function;
 import java.util.stream.Collectors;
 
 import org.infinispan.atomic.FineGrainedAtomicMap;
-import org.infinispan.container.MergeOnStore;
+import org.infinispan.container.impl.MergeOnStore;
 import org.infinispan.commands.remote.GetKeysInGroupCommand;
 import org.infinispan.functional.EntryView;
 import org.infinispan.commons.marshall.AdvancedExternalizer;