--- a/core/src/main/java/org/infinispan/commands/DataCommand.java
+++ b/core/src/main/java/org/infinispan/commands/DataCommand.java
@@ -8,6 +8,6 @@ package org.infinispan.commands;
  * @author Manik Surtani (<a href="mailto:manik@jboss.org">manik@jboss.org</a>)
  * @since 4.0
  */
-public interface DataCommand extends VisitableCommand, TopologyAffectedCommand, FlagAffectedCommand {
+public interface DataCommand extends VisitableCommand, TopologyAffectedCommand, FlagAffectedCommand, SegmentSpecificCommand {
    Object getKey();
 }