… core/src/main/java/org/infinispan/commands/functional/AbstractWriteKeyCommand.java
… core/src/main/java/org/infinispan/commands/functional/ReadOnlyKeyCommand.java
… core/src/main/java/org/infinispan/commands/functional/ReadWriteKeyCommand.java
… core/src/main/java/org/infinispan/commands/functional/ReadWriteKeyValueCommand.java
… core/src/main/java/org/infinispan/commands/functional/TxReadOnlyKeyCommand.java
… core/src/main/java/org/infinispan/commands/functional/WriteOnlyKeyCommand.java
… core/src/main/java/org/infinispan/commands/functional/WriteOnlyKeyValueCommand.java
… core/src/main/java/org/infinispan/commands/read/AbstractDataCommand.java
… core/src/main/java/org/infinispan/commands/read/GetAllCommand.java
… core/src/main/java/org/infinispan/commands/read/GetCacheEntryCommand.java
… core/src/main/java/org/infinispan/commands/read/GetKeyValueCommand.java
… core/src/main/java/org/infinispan/commands/read/KeySetCommand.java
… core/src/main/java/org/infinispan/commands/remote/ClusteredGetAllCommand.java
… core/src/main/java/org/infinispan/commands/remote/ClusteredGetCommand.java
… core/src/main/java/org/infinispan/commands/triangle/BackupWriteCommand.java
… core/src/main/java/org/infinispan/commands/triangle/SingleKeyBackupWriteCommand.java
… core/src/main/java/org/infinispan/commands/triangle/SingleKeyFunctionalBackupWriteCommand.java
… core/src/main/java/org/infinispan/commands/write/AbstractDataWriteCommand.java
… core/src/main/java/org/infinispan/commands/write/ApplyDeltaCommand.java
… core/src/main/java/org/infinispan/commands/write/ComputeCommand.java
… core/src/main/java/org/infinispan/commands/write/ComputeIfAbsentCommand.java
… core/src/main/java/org/infinispan/commands/write/EvictCommand.java
… core/src/main/java/org/infinispan/commands/write/PutKeyValueCommand.java
… core/src/main/java/org/infinispan/commands/write/RemoveCommand.java
… core/src/main/java/org/infinispan/commands/write/RemoveExpiredCommand.java
… core/src/main/java/org/infinispan/commands/write/ReplaceCommand.java
… core/src/main/java/org/infinispan/conflict/impl/DefaultConflictManager.java
… core/src/main/java/org/infinispan/container/AbstractSegmentedDataContainer.java
… core/src/main/java/org/infinispan/container/BoundedSegmentedDataContainer.java
… core/src/main/java/org/infinispan/container/DataContainer.java