https://issues.jboss.org/browse/ISPN-5451
https://issues.jboss.org/browse/ISPN-9212
https://issues.jboss.org/browse/ISPN-9245
https://issues.jboss.org/browse/ISPN-9246

Initial on heap implementation