@tristantarrant 2018-05-19
I'd like to see both performance and allocation details of this change.