@wburns 2018-05-21
Looks like it breaks a bunch of hibernate 2LC tests

https://ci.infinispan.org/job/Infinispan/job/PR-5997/3/#showFailuresLink