@wburns 2018-05-21
run performance tests please