@wburns 2018-05-21
Very quickly changing the https://github.com/infinispan/infinispan-benchmarks/tree/master/iteration to do a test for inserts with already precomputed keys and values I got the following:

```
SEGMENTED

Benchmark          (keyObjectSize)  (nodes)  (useIdentityCache)  (useStrings)  (valueObjectSize)   Mode  Cnt     Score    Error  Units
JMHBenchmarks.put               10        6                true         false               1000  thrpt   20  8839.924 ± 69.350  ops/s

Benchmark                                           (keyObjectSize)  (nodes)  (useIdentityCache)  (useStrings)  (valueObjectSize)   Mode  Cnt      Score       Error   Units
JMHBenchmarks.put                                                10        6                true         false               1000  thrpt   20   8587.319 ±    70.790   ops/s
JMHBenchmarks.put:·gc.alloc.rate                                 10        6                true         false               1000  thrpt   20    130.952 ±    22.624  MB/sec
JMHBenchmarks.put:·gc.alloc.rate.norm                            10        6                true         false               1000  thrpt   20  24037.155 ±  3990.484    B/op
JMHBenchmarks.put:·gc.churn.PS_Eden_Space                        10        6                true         false               1000  thrpt   20    133.564 ±   283.264  MB/sec
JMHBenchmarks.put:·gc.churn.PS_Eden_Space.norm                   10        6                true         false               1000  thrpt   20  24622.623 ± 52220.116    B/op
JMHBenchmarks.put:·gc.churn.PS_Survivor_Space                    10        6                true         false               1000  thrpt   20      0.031 ±     0.085  MB/sec
JMHBenchmarks.put:·gc.churn.PS_Survivor_Space.norm               10        6                true         false               1000  thrpt   20      5.809 ±    15.653    B/op
JMHBenchmarks.put:·gc.count                                      10        6                true         false               1000  thrpt   20      3.000              counts
JMHBenchmarks.put:·gc.time                                       10        6                true         false               1000  thrpt   20     16.000                  ms


MASTER

Benchmark          (keyObjectSize)  (nodes)  (useIdentityCache)  (useStrings)  (valueObjectSize)   Mode  Cnt     Score     Error  Units
JMHBenchmarks.put               10        6                true         false               1000  thrpt   20  8688.200 ± 171.789  ops/s


Benchmark                                           (keyObjectSize)  (nodes)  (useIdentityCache)  (useStrings)  (valueObjectSize)   Mode  Cnt      Score       Error   Units
JMHBenchmarks.put                                                10        6                true         false               1000  thrpt   20   8763.348 ±    53.413   ops/s
JMHBenchmarks.put:·gc.alloc.rate                                 10        6                true         false               1000  thrpt   20    133.793 ±    22.774  MB/sec
JMHBenchmarks.put:·gc.alloc.rate.norm                            10        6                true         false               1000  thrpt   20  24022.719 ±  3991.557    B/op
JMHBenchmarks.put:·gc.churn.PS_Eden_Space                        10        6                true         false               1000  thrpt   20    133.562 ±   283.259  MB/sec
JMHBenchmarks.put:·gc.churn.PS_Eden_Space.norm                   10        6                true         false               1000  thrpt   20  24132.248 ± 51179.819    B/op
JMHBenchmarks.put:·gc.churn.PS_Survivor_Space                    10        6                true         false               1000  thrpt   20      0.014 ±     0.053  MB/sec
JMHBenchmarks.put:·gc.churn.PS_Survivor_Space.norm               10        6                true         false               1000  thrpt   20      2.445 ±     9.496    B/op
JMHBenchmarks.put:·gc.count                                      10        6                true         false               1000  thrpt   20      3.000              counts
JMHBenchmarks.put:·gc.time                                       10        6                true         false               1000  thrpt   20     16.000                  ms

```

This was found with 2 runs of jmh for master and this branch. One without and one with `-prof gc` set. The results seem pretty consistent within each other.

From the code the only difference for a `Cache.put` and `Cache.get` should be one additional single segment lookup and array access for a given key.