@wburns 2018-05-21
I also did a 1 minute sampling of constant puts in the same environment with flight recorder and saw results that were quite similar

10 GCs for both averaging at 6.5ms each in duration.

The total TLAB allocations for SEGMENTED was 12.47 GB vs 12.22 GB (which is a 2% difference and well within sampling error). Looking at the allocations though both had 64% in byte[] alone.

I will run the tests with a LOCAL cache which show more meaningful numbers. The network overhead nullifies the data container itself.