@wburns 2018-05-21
With a DIST cache with 1 node I saw a 4% degradation, which is probably due to what I said earlier. I will see if there is any way I can trim that off a bit.

```
SEGMENTED

Benchmark          (keyObjectSize)  (nodes)  (useIdentityCache)  (useStrings)  (valueObjectSize)   Mode  Cnt       Score     Error  Units
JMHBenchmarks.put               10        1                true          true               1000  thrpt  100  148785.318 ± 754.433  ops/s

MASTER
Benchmark          (keyObjectSize)  (nodes)  (useIdentityCache)  (useStrings)  (valueObjectSize)   Mode  Cnt       Score      Error  Units
JMHBenchmarks.put               10        1                true          true               1000  thrpt  100  154839.369 ± 1114.430  ops/s
```