@wburns 2018-05-21
Unfortunately all of my previous runs have been invalidated as I didn't realize I had a store configured, oops!

A new run with single node DIST showed SEGMENTED actually faster in the run, so my guess is that they should be pretty much the same. I have to verify allocation rates still and radargun perf test should be done soon.

```
SEGMENT

Benchmark          (keyObjectSize)  (nodes)  (useIdentityCache)  (useStrings)  (valueObjectSize)   Mode  Cnt        Score      Error  Units
JMHBenchmarks.put               10        1                true          true               1000  thrpt   20  1170006.851 ± 9396.602  ops/s


MASTER

Benchmark          (keyObjectSize)  (nodes)  (useIdentityCache)  (useStrings)  (valueObjectSize)   Mode  Cnt        Score       Error  Units
JMHBenchmarks.put               10        1                true          true               1000  thrpt   20  1168134.261 ± 37600.542  ops/s
```