@wburns 2018-05-21
Performance job shows no major regressions, however I will see if I can tweak a few things.